RepoOrganization

+++ mobile_robotics_slam/MapGenerator/OnlineMap.py
@@ -7,19 +7,27 @@ import os
 
 class DynamicMapUpdater:
     def __init__(self):
+        print("DYNAMIC MAP STARTED")
         self.data_queue = multiprocessing.Queue()
         self.process = multiprocessing.Process(target=self._update_map, args=(self.data_queue,))
         self.process.daemon = True   # Ensures the thread closes with the main program
-        self.update_interval = 0.1  # Update every 100ms
-        self.frames_dir = "frames"  # Directory to save frames
-        self.gif_path = "dynamic_map_no_icp.gif"  # Path for the final GIF
+        self.update_interval = 2  # Update every 100ms
+        
+        path = __file__
+        file_location_subfolders = 3 #Number of folder to go up to reach root of package
+        for _ in range(file_location_subfolders):
+            path = os.path.dirname(path)
 
+        
+        self.frames_dir = os.path.join(path, "frames")
+        
         # Create frames directory if it doesn't exist
         if not os.path.exists(self.frames_dir):
             os.makedirs(self.frames_dir)
         else: # Clear existing frames
             for f in os.listdir(self.frames_dir):
-                os.remove(os.path.join(self.frames_dir, f))
+                if f.endswith(".png"):
+                    os.remove(os.path.join(self.frames_dir, f))
 
     def start(self):
         self.process.start()
@@ -32,7 +40,7 @@ class DynamicMapUpdater:
 
 
     def add_data(self, poses, landmarks, points):
-        while self.data_queue.qsize() >= 4:
+        while self.data_queue.qsize() >= 2:
             try:
                 self.data_queue.get_nowait()  # Remove the oldest data
             except Empty:
@@ -43,15 +51,16 @@ class DynamicMapUpdater:
         
 
     def _update_map(self, data_queue):
-        plt.ion()  # Enable interactive mode
+        #plt.ion()  # Enable interactive mode
         fig, ax = plt.subplots()
         
-        angles = np.linspace(-np.pi, np.pi, 4000)
+        
         frame_count = 0
         while True:
             try:
                 # Retrieve latest data
                 poses, landmarks, ranges = data_queue.get(timeout=self.update_interval)
+                
                 ax.clear()
                 print("Updating Map")
 
@@ -66,9 +75,10 @@ class DynamicMapUpdater:
                     poses = np.array(poses)
                     ax.plot(poses[:, 0], poses[:, 1], "orange", label='Optimized Trajectory')
                     for pose, range in zip(poses, ranges):
+                        angles = np.linspace(-np.pi, np.pi, len(range))
                         x = pose[0] + range * np.cos(angles + pose[2])
                         y = pose[1] + range * np.sin(angles + pose[2])
-                        x, y = x[range < 9], y[range < 9]
+                        x, y = x[range < 5], y[range < 5]
                         map.extend(np.vstack((x, y)).T)
                     map = np.array(map)
                     ax.scatter(map[:, 0], map[:, 1], c='g', s=1)
@@ -78,7 +88,7 @@ class DynamicMapUpdater:
                 # Plot landmarks
                 if landmarks is not None and len(landmarks) > 0:
                     landmarks = np.array(landmarks)
-                    ax.scatter(landmarks[:, 0], landmarks[:, 1], c="r", label="Landmarks")
+                    ax.scatter(landmarks[:, 0], landmarks[:, 1], c="r", label="Corners", s=4)
 
                 ax.set_title("Dynamic Map")
                 ax.set_aspect('equal')
@@ -97,19 +107,3 @@ class DynamicMapUpdater:
                 print(e)
                 # Handle queue timeout or other exceptions
                 continue
-
-        plt.ioff()
-        plt.show()
-
-    def _generate_gif(self):
-        # Create GIF from saved frames
-        frames = []
-        frame_files = sorted([os.path.join(self.frames_dir, f) for f in os.listdir(self.frames_dir) if f.endswith(".png")])
-        for frame_file in frame_files:
-            frames.append(imageio.v3.imread(frame_file))
-
-        if frames:
-            imageio.mimsave(self.gif_path, frames, duration=self.update_interval)
-            print(f"GIF saved to {self.gif_path}")
-        else:
-            print("No frames to create a GIF.")
